transferred button utility functions to new class
